chore: rename HeroSection
closes #33
